moved files to appropriate packages, and renamed
