+++ src/main/java/com/mamiglia/utils/DisplayLogger.java
@@ -1,4 +1,4 @@
-package Utils;
+package com.mamiglia.utils;
 
 import java.io.File;
 import java.io.IOException;
